+++ R/ergm-tidiers.R
@@ -8,19 +8,17 @@
 #' @param x An `ergm` object returned from a call to [ergm::ergm()].
 #' @template param_confint
 #' @template param_exponentiate
-#' @template param_quick
 #' @param ... Additional arguments to pass to [ergm::summary()].
 #'   **Cautionary note**: Mispecified arguments may be silently ignored.
 #' 
-#' @evalRd return_tidy(
-#'   "term", 
-#'   "estimate", 
-#'   "std.error", 
-#'   "mcmc.error", 
-#'   "p.value",
-#'   "conf.low",
-#'   "conf.high"
-#' )
+#'
+#' @return A [tibble::tibble] with one row for each coefficient in the
+#'   exponential random graph model, with columns:
+#'   \item{term}{The term in the model being estimated and tested}
+#'   \item{estimate}{The estimated coefficient}
+#'   \item{std.error}{The standard error}
+#'   \item{mcmc.error}{The MCMC error}
+#'   \item{p.value}{The two-sided p-value}
 #'
 #' @examples
 #' 
@@ -55,22 +53,41 @@
 #' @seealso [tidy()], [ergm::ergm()], [ergm::control.ergm()], 
 #'   [ergm::summary()]
 #' @family ergm tidiers
-tidy.ergm <- function(x, conf.int = FALSE, conf.level = .95,
-                      exponentiate = FALSE, quick = FALSE, ...) {
-  if (quick) {
-    co <- x$coef
-    ret <- tibble(term = names(co), estimate = unname(co))
-    return(process_ergm(ret, conf.int = FALSE, exponentiate = exponentiate))
-  }
-  co <- ergm:::summary.ergm(x, ...)$coefs
-
-  nn <- c("estimate", "std.error", "mcmc.error", "p.value")
-  ret <- fix_data_frame(co, nn[1:ncol(co)])
+tidy.ergm <- function(x, conf.int = FALSE, conf.level = 0.95,
+                      exponentiate = FALSE,  ...) {
   
-  process_ergm(ret, x,
-    conf.int = conf.int, conf.level = conf.level,
-    exponentiate = exponentiate
-  )
+  ret <- summary(x, ...)$coefs %>% 
+    tibble::rownames_to_column() %>% 
+    rename2(
+      term = "rowname",
+      estimate = "Estimate", 
+      std.error = "Std. Error",
+      mcmc.error = "MCMC %",
+      statistic = "z value",
+      p.value = "Pr(>|z|)"
+    )
+  
+  if (conf.int) {
+    z <- stats::qnorm(1 - (1 - conf.level) / 2)
+    ret$conf.low <- ret$estimate - z * ret$std.error
+    ret$conf.high <- ret$estimate + z * ret$std.error
+  }
+  
+  if (exponentiate) {
+    
+    if (is.null(x$glm) ||
+        (x$glm$family$link != "logit" && x$glm$family$link != "log"))
+      warning("Exponentiating but model didn't use log or logit link.")
+    
+    ret$estimate <- exp(ret$estimate)
+    
+    if (conf.int) {
+      ret$conf.low <- exp(ret$conf.low)
+      ret$conf.high <- exp(ret$conf.high)
+    }
+  }
+  
+  as_tibble(ret)
 }
 
 #' @templateVar class ergm
@@ -83,88 +100,61 @@ tidy.ergm <- function(x, conf.int = FALSE, conf.level = .95,
 #' @param mcmc Logical indicating whether or not to report MCMC interval, 
 #'   burn-in and sample size used to estimate the model. Defaults to `FALSE`.
 #'
-#' @evalRd return_glance(
-#'   independence = "Whether the model assumed dyadic independence",
-#'   "iterations",
-#'   "logLik",
-#'   "AIC",
-#'   "BIC",
-#'   "null.deviance",
-#'   "df.null",
-#'   "residual.deviance",
-#'   "df.residual",
-#'   "MCMC.interval",
-#'   "MCMC.burnin",
-#'   "MCMC.samplesize"
-#' )
+#' @return `glance.ergm` returns a one-row data.frame with the columns
+#'   \item{independence}{Whether the model assumed dyadic independence}
+#'   \item{iterations}{The number of MCMLE iterations performed before convergence}
+#'   \item{logLik}{If applicable, the log-likelihood associated with the model}
+#'   \item{AIC}{The Akaike Information Criterion}
+#'   \item{BIC}{The Bayesian Information Criterion}
+#'
+#' If `deviance = TRUE`, and if the model supports it, the
+#' data frame will also contain the columns
+#'   \item{null.deviance}{The null deviance of the model}
+#'   \item{df.null}{The degrees of freedom of the null deviance}
+#'   \item{residual.deviance}{The residual deviance of the model}
+#'   \item{df.residual}{The degrees of freedom of the residual deviance}
 #'
 #' @export
 #' @seealso [glance()], [ergm::ergm()], [ergm::summary.ergm()]
 #' @family ergm tidiers
 glance.ergm <- function(x, deviance = FALSE, mcmc = FALSE, ...) {
-  # will show appropriate warnings about standard errors, pseudolikelihood etc.
-  s <- ergm:::summary.ergm(x, ...)
-  # dyadic (in)dependence and number of MCMLE iterations
-  ret <- tibble(independence = s$independence, iterations = x$iterations)
-  # log-likelihood
-  ret$logLik <- tryCatch(as.numeric(ergm:::logLik.ergm(x)), error = function(e) NULL)
-  # null and residual deviance
+  
+  s <- summary(x, ...)  # produces lots of messages
+  
+  ret <- tibble(
+    independence = s$independence,
+    iterations = x$iterations,
+    logLik = as.numeric(logLik(x))
+  )
+  
   if (deviance & !is.null(ret$logLik)) {
-    # thanks to Pavel N. Krivitsky (@krivit)
-    # https://github.com/tidymodels/broom/issues/567
-    if (packageVersion("ergm") < "3.10") {
-      dyads <- sum(as.rlebdm(x$constrained, x$constrained.obs, which = "informative"))
+    
+    # see #567 for details on the following
+    
+    if (utils::packageVersion("ergm") < "3.10") {
+      dyads <- sum(
+        ergm::as.rlebdm(x$constrained, x$constrained.obs, which = "informative")
+      )
     } else {
       dyads <- stats::nobs(x)
     }
-
-    ret$null.deviance <- ergm:::logLikNull(x)
-    ret$null.deviance <- ifelse(is.na(ret$null.deviance), 0, -2 * ret$null.deviance)
+    
+    lln <- ergm::logLikNull(x)
+    ret$null.deviance <- if (is.na(lln)) 0 else -2 * lln
     ret$df.null <- dyads
 
     ret$residual.deviance <- -2 * ret$logLik
-    ret$df.residual <- dyads - length(x$coef)
-  }
-  ret$AIC <- tryCatch(stats::AIC(x), error = function(e) NULL)
-  ret$BIC <- tryCatch(stats::BIC(x), error = function(e) NULL)
-
-  if (mcmc) {
-    ret <- cbind(ret, data.frame(
-      MCMC.interval = x$control$MCMC.interval,
-      MCMC.burnin = x$control$MCMC.burnin,
-      MCMC.samplesize = x$control$MCMC.samplesize
-    ))
+    ret$df.residual <- dyads - length(x$coefs)
   }
   
-  as_tibble(ret)
-}
+  ret$AIC <- stats::AIC(x)
+  ret$BIC <- stats::BIC(x)
 
-# helper function
-process_ergm <- function(ret, x, conf.int = FALSE, conf.level = .95,
-                         exponentiate = FALSE) {
-  if (exponentiate) {
-    # save transformation function for use on confidence interval
-    if (is.null(x$glm) ||
-      (x$glm$family$link != "logit" && x$glm$family$link != "log")) {
-      warning(paste(
-        "Exponentiating coefficients, but model did not use",
-        "a log or logit link function"
-      ))
-    }
-    trans <- exp
-  } else {
-    trans <- identity
+  if (mcmc) {
+    ret$MCMC.interval <- x$control$MCMC.interval
+    ret$MCMC.burnin <- x$control$MCMC.burnin
+    ret$MCMC.samplesize <- x$control$MCMC.samplesize
   }
-
-  if (conf.int) {
-    z <- stats::qnorm(1 - (1 - conf.level) / 2)
-    CI <- cbind(
-      conf.low = ret$estimate - z * ret$std.error,
-      conf.high = ret$estimate + z * ret$std.error
-    )
-    ret <- cbind(ret, trans(unrowname(CI)))
-  }
-  ret$estimate <- trans(ret$estimate)
-
-  as_tibble(ret)
+  
+  ret
 }