+++ NEWS.md
@@ -1,4 +1,4 @@
-# broom 0.5.2.9000
+# broom 0.5.2.9001
 (To be released as 0.7.0)
 
 ## Breaking changes
@@ -11,8 +11,12 @@ changes in this version of `broom`. We list them below-
     `AIC`, `BIC, deviance`, `nobs`. This is in response to conversation that
     took place in #212. Note that `tidy.aov` can be used to get numerator and
     denominator degrees of freedom.
+  - Augment method for `factanal` objects now returns a tibble with name
+    pattern `.fs` (e.g., `.fs1`, `.fs2`, `.fs3`, etc.), instead of `factor`
+    (e.g., `factor1`, `factor2`, `factor3`, etc.) (#650).
     
-
+  - We have removed all support for the `quick` argument in `tidy()` methods. TODO: explain why, and discuss alternatives.
+    
 ## Changes to `augment()`
 
 have overhauled `augment()` for general consistency improvements (hopefully,
@@ -68,7 +72,16 @@ TODO: sort out what happens to `glance.aov()`
 
 ## New tidiers, features and bugfixes
 
- 
+- Added `tidy()` and `glance()` methods for `speedglm` objects from the `speedglm` package
+
+- Added tidier for `epiR::epi.2by2` (#711)
+
+- Make `.fitted` values respect `type.predict` argument of `augment.clm()`. (#617)
+
+- Return factor rather than numeric class predictions in `.fitted` of `augment.polr()`. (#619)
+
+- `ordinal` tidier rewrite
+
 - Added tidiers for `rma` objects from the `metafor` package (#674, @malcolmbarrett, @softloud)
 
 - Added support for `tidy.lavaan()` to take `quick = TRUE`. (#628)
@@ -83,6 +96,15 @@ TODO: sort out what happens to `glance.aov()`
 - Added `tidy.regsubsets()` for best subsets linear regression from the `leaps` package
 
 - Added method `tidy.lm.beta()` to tidy `lm.beta` class models (#545 by @mattle24)
+
+- Add feature for glance.biglm to return df.residual
+
+- Patch bug in glance.lavaan (#577)
+
+- Added method `tidy.systemfit()` to tidy `systemfit` class models (by @jaspercooper)
+
+- Added tidiers for `drc::drm` models (#574 by @edild)
+
 - `tidy.prcomp()` parameter `matrix` gained new options `"scores"`, `"loadings"`, and `"eigenvalues"` (#557 by @GegznaV)
 
 - `tidy.kmeans()` now uses the names of the input variables in the output by